add Spring Security and endpoint to change email

+++ src/main/java/pl/maryana/conference/service/implementation/ReservationServiceImpl.java
@@ -3,6 +3,7 @@ package pl.maryana.conference.service.implementation;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import pl.maryana.conference.exception.*;
 import pl.maryana.conference.model.Lecture;
 import pl.maryana.conference.model.Reservation;
@@ -13,6 +14,7 @@ import pl.maryana.conference.service.MailService;
 import pl.maryana.conference.service.ReservationService;
 import pl.maryana.conference.service.UserService;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -24,16 +26,19 @@ public class ReservationServiceImpl implements ReservationService {
     private final UserService userService;
     private final LectureService lectureService;
     private final MailService mailService;
+    private final TimeService timeService;
 
     @Autowired
-    public ReservationServiceImpl(ReservationRepository reservationRepository, UserService userService, LectureService lectureService, MailService mailService) {
+    public ReservationServiceImpl(TimeService timeService, ReservationRepository reservationRepository, UserService userService, LectureService lectureService, MailService mailService) {
         this.reservationRepository = reservationRepository;
         this.userService = userService;
         this.lectureService = lectureService;
         this.mailService = mailService;
+        this.timeService = timeService;
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Reservation> findAllByLogin(String login) {
 
         Optional<User> user = userService.findByLogin(login);
@@ -51,22 +56,21 @@ public class ReservationServiceImpl implements ReservationService {
     @Override
     public Reservation reserve(String login, String email, long lectureId) {
 
+        Lecture lecture = lectureService.findById(lectureId).orElseThrow(() -> new LectureNotFound("Lecture with id [" + lectureId +"] not found"));
+        LocalDateTime lectureStartDateTime = lecture.getStartDateTime();
+
+        if(timeService.isExpired(lectureStartDateTime))
+            throw new LectureReservationExpiredException("Lecture has already started. You cannot make reservation");
+
         if(reservationRepository.countAllByLectureId(lectureId) == 5)
             throw new LimitOfReservations("Limit of reservation of lecture with id [" + lectureId + "] reached");
 
 
-        Optional<User> user = userService.findByLogin(login);
-        if(user.isPresent()){
+        if(reservationRepository.findByLectureIdAndUserLogin(lectureId, login).isEmpty()){
 
-            if(reservationRepository.findByUserAndLectureId(user.get(), lectureId).isPresent())
-                throw new LoginIsAlreadyTaken("Podany login jest już zajęty");
-
-            else{
                 Reservation reservation = new Reservation();
                 reservation.setLectureId(lectureId);
-                reservation.setUser(user.get());
-
-                Lecture lecture = lectureService.findById(lectureId).orElseThrow(() -> new LectureNotFound("Lecture with id [" + lectureId +"] not found"));
+                reservation.setUser(userService.findByLogin(login).orElseThrow(() -> new UserNotFound("User [" + login +"] not found")));
 
                 reservation.setLecture(lecture);
 
@@ -76,45 +80,23 @@ public class ReservationServiceImpl implements ReservationService {
                 log.info("Reservation to "  + lectureService.findById(lectureId) + " of user " + login + " was saved");
 
                 return reservation;
-            }
-        }
-
-        User newUser = new User();
-        newUser.setLogin(login);
-        newUser.setEmail(email);
-
-        Reservation reservation = new Reservation();
-        reservation.setLectureId(lectureId);
-        newUser.addReservation(reservation);
-
-        Lecture lecture = lectureService.findById(lectureId).orElseThrow(() -> new LectureNotFound("Lecture with id [" + lectureId +"] not found"));
-        reservation.setLecture(lecture);
-
-        mailService.sendEmail(email, login, lecture);
-
-        userService.save(newUser);
-        log.info("New user with login " + login +" was saved and his reservation to "  + lectureService.findById(lectureId) + " was saved");
-
-        return reservation;
+     }
+        log.info("User [" + login + "] tries make reservation for the lecture twice");
+        throw new DuplicateReservationException("User have already made a reservation for this lecture");
     }
 
     @Override
-    public void cancelReservation(long reservationId) {
+    public void cancelReservation(long reservationId, String login) {
+
         Reservation reservation = reservationRepository.findById(reservationId).orElseThrow(() ->
                 new ReservationNotFound("Reservation with id [" + reservationId + "] not found"));
+
+        if(reservationRepository.findByIdAndUserLogin(reservationId, login).isEmpty()) {
+            log.info("User [" + login +"] try do cancel not his reservation [" + reservationId +"]");
+            throw new UnauthorizedCancellationException("User [" + login + "] doesn't have access to cancel reservation [" + reservationId + "]");
+        }
+
         reservationRepository.delete(reservation);
     }
 
-    @Override
-    public Reservation findById(long reservationId) {
-
-        Reservation reservation = reservationRepository.findById(reservationId).orElseThrow(() ->
-                new ReservationNotFound("Reservation with id [" + reservationId + "] not found"));
-
-        Lecture lecture = lectureService.findById(reservation.getLectureId()).orElseThrow(() ->
-                new LectureNotFound("Lecture with id [" + reservation.getLectureId() +"] not found"));
-
-        reservation.setLecture(lecture);
-        return reservation ;
-    }
 }
